==> 09_WFP_Chad_MAD_data.Rmd <==
---
title: "WFP datasets wrangling - Chad : Régime alimentaire minimum acceptable"
author: "Aboubacar HEMA"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    number_sections: true
    code_folding: show
    keep_md: true
---



```{r 09setup, include=FALSE, results = 'hide'}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE,
  results = "hide"
)
```

# MAD indicator

```{r}
library(haven)
library(labelled) # for general functions to work with labelled data
library(tidyverse) # general wrangling
library(dplyr)
library(Hmisc)
library(gtsummary) # to demonstrate automatic use of variable labels in summary tables
library(readxl)
library(foreign)
library(sjPlot)
library(sjmisc)
library(sjlabelled) # for example efc data set with variable labels
library(stringr)
```




```{r}
rm(list = ls())
```


```{r environment settings}

dir_input_data = "C:/Users/AHema/OneDrive - CGIAR/Desktop/WFP Resilience dataset/data/input_data/Chad"
dir_output_data = "C:/Users/AHema/OneDrive - CGIAR/Desktop/WFP Resilience dataset/data/output_data/Chad"
```

```{r read harmonization excel file}

Chad_Harmonization_variables <- read_excel(paste0(dir_input_data,"/Chad_Harmonization.xlsx"), 
    sheet = "variables_harmonization")
#View(Chad_Harmonization_variables)

Chad_Harmonization_description <- read_excel(paste0(dir_input_data,"/Chad_Harmonization.xlsx"), 
    sheet = "description")
#View(Chad_Harmonization_description)
```

```{r}

lst_data = Chad_Harmonization_description$Data
lst_test = Chad_Harmonization_description$Name

for(i in 1:length(lst_data)) {                              # Head of for-loop
  assign(lst_test[i],                                   # Read and store data frames
         read_sav(paste0(dir_input_data,"/",lst_data[i])))
}

```

## Data consistency  Check

```{r eval=FALSE, include=FALSE}
Chad_baseline_2018$ID <- Chad_baseline_2018$`@_id`
Chad_ea_2019$ID <- Chad_ea_2019$`@_id`
Chad_ea_2020$ID <- Chad_ea_2020$Identifiant
Chad_pdm_2020$ID <- Chad_pdm_2020$Identifiant
Chad_ea_2021$ID <- Chad_ea_2021$`@_id`
Chad_pdm_2021$ID <- Chad_pdm_2021$Identifiant
Chad_ea_2022$ID <- Chad_ea_2022$`@_id`
Chad_pdm_2023$ID <- Chad_pdm_2023$`@_id`
Chad_ea_2023$ID <- Chad_ea_2023$`@_id`

Chad_baseline_2018$RESPConsent <- Chad_baseline_2018$resultat_enquete
Chad_ea_2019$RESPConsent <- Chad_ea_2019$`@2.1.2Linterviewéavalidélanotedeconsentement`
Chad_ea_2020$RESPConsent <- Chad_ea_2020$q212_notevld
#Chad_pdm_2020$RESPConsent <- Chad_pdm_2020
#Chad_ea_2021$RESPConsent <- Chad_ea_2021$`@_RESPConsent`
Chad_pdm_2021$RESPConsent <- Chad_pdm_2021$q212_notevld
Chad_ea_2022$RESPConsent <- Chad_ea_2022$RESPConsent
Chad_pdm_2022$RESPConsent <- Chad_pdm_2022$RESPConsent
Chad_pdm_2023$RESPConsent <- Chad_pdm_2023$RESPConsent
Chad_ea_2023$RESPConsent <- Chad_ea_2023$RESPConsent

Chad_baseline_2018$ADMIN1Name <- Chad_baseline_2018$region
Chad_ea_2019$ADMIN1Name <- Chad_ea_2019$`@1.02NomdelaProvince`
Chad_ea_2020$ADMIN1Name <- Chad_ea_2020$province
Chad_pdm_2020$ADMIN1Name <- Chad_pdm_2020$q4_provinces
Chad_ea_2021$ADMIN1Name <- Chad_ea_2021$ADMIN1Name
Chad_pdm_2021$ADMIN1Name <- Chad_pdm_2021$province
Chad_ea_2022$ADMIN1Name <- Chad_ea_2022$ADMIN1Name
Chad_pdm_2022$ADMIN1Name <- Chad_pdm_2022$ADMIN1Name
Chad_pdm_2023$ADMIN1Name <- Chad_pdm_2023$ADMIN1Name
Chad_ea_2023$ADMIN1Name <- Chad_ea_2023$ADMIN1Name
```





## Drop duplicated observations

```{r eval=FALSE, include=FALSE}
Chad_baseline_2018 <- Chad_baseline_2018 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_ea_2019 <- Chad_ea_2019 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_ea_2020 <- Chad_ea_2020 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_ea_2021 <- Chad_ea_2021 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_ea_2022 <- Chad_ea_2022 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_ea_2023 <- Chad_ea_2023 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_pdm_2020 <- Chad_pdm_2020 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_pdm_2021 <- Chad_pdm_2021 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
Chad_pdm_2022 <- Chad_pdm_2022 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))

Chad_pdm_2023 <- Chad_pdm_2023 %>% dplyr::distinct() %>% dplyr::filter(!is.na(ADMIN1Name))
```

## Consent check

```{r eval=FALSE, include=FALSE}
Chad_baseline_2018 <- Chad_baseline_2018  %>% filter(RESPConsent == 1)
Chad_ea_2019 <- Chad_ea_2019  %>% filter(RESPConsent == "Oui")
Chad_ea_2020 <- Chad_ea_2020  %>% filter(RESPConsent == 1)
#Chad_ea_2021 <- Chad_ea_2021  %>% filter(RESPConsent =="")
Chad_ea_2022 <- Chad_ea_2022 %>% filter(RESPConsent == 2)
Chad_ea_2023 <- Chad_ea_2023 %>% filter(RESPConsent == 1)
#Chad_pdm_2020 <- Chad_pdm_2020  %>% filter(RESPConsent =="")
Chad_pdm_2021 <- Chad_pdm_2021  %>% filter(RESPConsent == 1)
Chad_pdm_2022 <- Chad_pdm_2022  %>% filter(RESPConsent == 2)
Chad_pdm_2023 <- Chad_pdm_2023  %>% filter(RESPConsent == 1)
```

## ID Check 

```{r eval=FALSE, include=FALSE}
Chad_baseline_2018 <- Chad_baseline_2018 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID))
Chad_ea_2019 <- Chad_ea_2019 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID))
Chad_ea_2020 <- Chad_ea_2020 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID))
Chad_ea_2021 <- Chad_ea_2021 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID))
Chad_ea_2022 <- Chad_ea_2022 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID)) 

Chad_ea_2023 <- Chad_ea_2023 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID)) 

Chad_pdm_2020 <- Chad_pdm_2020 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID)) 
Chad_pdm_2021 <- Chad_pdm_2021 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID)) 
#Chad_pdm_2022 <- Chad_pdm_2022 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID)) 

Chad_pdm_2023 <- Chad_pdm_2023 %>% dplyr::distinct(ID,.keep_all = TRUE) %>% dplyr::filter(!is.na(ID))
Chad_pdm_2022$ID <- 1:nrow(Chad_pdm_2022)
Chad_pdm_2022 <- labelled::to_factor(Chad_pdm_2022)

```





```{r}
path = "C:/Users/AHema/OneDrive - CGIAR/Desktop/WFP Resilience dataset/"

```

```{r}
MAD_ea_2021 <- read_sav(paste0(path,"data/input_data/Chad/MAD_MDDW data/consent_PCMADRepeat623_Decembre_2021.sav"))
MAD_ea_2022 <- read_sav(paste0(path,"data/input_data/Chad/MAD_MDDW data/consent_PCMADRepeat623_Decembre_2022.sav"))
MAD_ea_2023 <- read_sav(paste0(path,"data/input_data/Chad/MAD_MDDW data/consent_PCMADRepeat623_Decembre_2023.sav"))
MAD_pdm_2022 <- read_sav(paste0(path,"data/input_data/Chad/MAD_MDDW data/consent_PCMADRepeat623_Juillet_2022.sav"))
MAD_pdm_2023 <- read_sav(paste0(path,"data/input_data/Chad/MAD_MDDW data/consent_PCMADRepeat623_Juillet_2023.sav"))

```

```{r}
Chad_pdm_2022$ID = 1:nrow(Chad_pdm_2022)
```



## EA 2021

```{r}

df_ea_2021 = Chad_ea_2021 %>% 
  dplyr::select(`@_id`,`@_index`)

names(df_ea_2021) <- c("ID","Key_with_subset_data")

MAD_ea_2021$Key_with_subset_data <- MAD_ea_2021$`@_parent_index`
MAD_ea_2021 = dplyr::left_join(df_ea_2021, MAD_ea_2021, by = "Key_with_subset_data")

MAD_ea_2021 = MAD_ea_2021 %>% 
  dplyr::select(!starts_with("@"))

MAD_ea_2021 = MAD_ea_2021 %>% 
  dplyr::filter(!is.na(PCMAD_childnumber))
```


## EA 2022

```{r}
df_ea_2022 = Chad_ea_2022 %>% 
  dplyr::select(`@_id`,`@_index`)

names(df_ea_2022) <- c("ID","Key_with_subset_data")

MAD_ea_2022$Key_with_subset_data <- MAD_ea_2022$`@_parent_index`
MAD_ea_2022 = dplyr::left_join(df_ea_2022, MAD_ea_2022, by = "Key_with_subset_data")

MAD_ea_2022 = MAD_ea_2022 %>% 
  dplyr::select(!starts_with("@"))

MAD_ea_2022 = MAD_ea_2022 %>% 
  dplyr::filter(!is.na(PCMAD_childnumber))
```



## EA 2023

```{r}
df_ea_2023 = Chad_ea_2023 %>% 
  dplyr::select(`@_id`,`@_index`)

names(df_ea_2023) <- c("ID","Key_with_subset_data")

MAD_ea_2023$Key_with_subset_data <- MAD_ea_2023$`@_parent_index`
MAD_ea_2023 = dplyr::left_join(df_ea_2023, MAD_ea_2023, by = "Key_with_subset_data")

MAD_ea_2023 = MAD_ea_2023 %>% 
  dplyr::select(!starts_with("@"))

MAD_ea_2023 = MAD_ea_2023 %>% 
  dplyr::filter(!is.na(PCMAD_childnumber))

MAD_ea_2023 <- MAD_ea_2023 %>%
  mutate(across(MAD_sex, ~ifelse(.==0, 1, 2)))


# Labels mapping
label_map <- c(
  Femme = 1,
  Homme = 2
)

# Apply labels using mutate() and across()
MAD_ea_2023 <- MAD_ea_2023 %>%
  mutate(across(MAD_sex,
                ~ labelled::labelled(., label_map)))



MAD_ea_2023 <- MAD_ea_2023 %>%
  mutate(across(c(EverBreastF,PCIYCBreastF,PCMADStapCer,PCMADVegOrg,PCMADStapRoo,PCMADVegGre,PCMADFruitOrg,PCMADVegFruitOth,PCMADPrMeatO,PCMADPrMeatF,PCMADPrEgg,PCMADPrFish,PCMADPulse,PCMADDairy,PCMADFatRpalm,PCMADSnfChild,PCMADSnfPowd,PCMADSnfLns), ~ifelse(is.na(.),NA,ifelse(.==0, 1, ifelse(.==1,2,3)))))
# Labels mapping
label_map <- c(
  Non = 1,
  Oui = 2,
  `Ne sait pas` = 3
)

# Apply labels using mutate() and across()
MAD_ea_2023 <- MAD_ea_2023 %>%
  mutate(across(c(EverBreastF,PCIYCBreastF,PCMADStapCer,PCMADVegOrg,PCMADStapRoo,PCMADVegGre,PCMADFruitOrg,PCMADVegFruitOth,PCMADPrMeatO,PCMADPrMeatF,PCMADPrEgg,PCMADPrFish,PCMADPulse,PCMADDairy,PCMADFatRpalm,PCMADSnfChild,PCMADSnfPowd,PCMADSnfLns),
                ~ labelled::labelled(., label_map)))
```



## PDM 2022

```{r}

df_pdm_2022 = Chad_pdm_2022 %>% 
  dplyr::select("ID",`@_index`)

names(df_pdm_2022) = c("ID","Key_with_subset_data")

MAD_pdm_2022$Key_with_subset_data <- MAD_pdm_2022$`@_parent_index`
MAD_pdm_2022 = dplyr::left_join(df_pdm_2022, MAD_pdm_2022, by = "Key_with_subset_data")

MAD_pdm_2022 = MAD_pdm_2022 %>% 
  dplyr::select(!starts_with("@"))

MAD_pdm_2022 = MAD_pdm_2022 %>% 
  dplyr::filter(!is.na(PCMAD_childnumber))
```



## PDM 2023


```{r}
df_pdm_2023 = Chad_pdm_2023 %>% 
  dplyr::select(`@_id`,`@_index`)

names(df_pdm_2023) <- c("ID","Key_with_subset_data")
MAD_pdm_2023$Key_with_subset_data <- MAD_pdm_2023$`@_parent_index`
MAD_pdm_2023 = dplyr::left_join(df_pdm_2023, MAD_pdm_2023, by = "Key_with_subset_data")

MAD_pdm_2023 = MAD_pdm_2023 %>% 
  dplyr::select(!starts_with("@"))

MAD_pdm_2023 = MAD_pdm_2023 %>% 
  dplyr::filter(!is.na(PCMAD_childnumber))
```


## Régime alimentaire minimum acceptable (MAD)

## Export all data

```{r}
dir_output_data = "C:/Users/AHema/OneDrive - CGIAR/Desktop/WFP Resilience dataset/data/output_data/Chad"
```


```{r eval=FALSE, include=FALSE}
write_sav(MAD_ea_2021, paste0(dir_output_data,"/MAD_ea_2021.sav"))
write_dta(MAD_ea_2022, paste0(dir_output_data,"/MAD_ea_2022.dta"))
write_sav(MAD_ea_2023, paste0(dir_output_data,"/MAD_ea_2023.sav"))
write_dta(MAD_pdm_2022, paste0(dir_output_data,"/MAD_pdm_2022.dta"))
write_sav(MAD_pdm_2023, paste0(dir_output_data,"/MAD_pdm_2023.sav"))

```

```{r}
WFP_Chad_MAD<-plyr::rbind.fill(MAD_ea_2021,
                                MAD_ea_2022,
                                MAD_ea_2023,
                                MAD_pdm_2022,
                                MAD_pdm_2023)


WFP_Chad_MAD$ID <- as.character(WFP_Chad_MAD$ID)
WFP_Chad_MAD = WFP_Chad_MAD %>% 
  dplyr::select(-c("Key_with_subset_data","MAD_name","MAD_dob","PCMAD_childnumber"))


WFP_Chad_MAD <- WFP_Chad_MAD %>%
  mutate(across(MAD_sex, ~ifelse(is.na(.),NA,ifelse(.==1, 0, 1))))


# Labels mapping
label_map <- c(
  Femme = 0,
  Homme = 1
)

# Apply labels using mutate() and across()
WFP_Chad_MAD <- WFP_Chad_MAD %>%
  mutate(across(MAD_sex,
                ~ labelled::labelled(., label_map)))


WFP_Chad_MAD <- WFP_Chad_MAD %>%
  mutate(across(c(EverBreastF,PCIYCBreastF,PCMADStapCer,PCMADVegOrg,PCMADStapRoo,PCMADVegGre,PCMADFruitOrg,PCMADVegFruitOth,PCMADPrMeatO,PCMADPrMeatF,PCMADPrEgg,PCMADPrFish,PCMADPulse,PCMADDairy,PCMADFatRpalm,PCMADSnfChild,PCMADSnfPowd,PCMADSnfLns), ~ifelse(is.na(.),NA,ifelse(.==1,0, ifelse(.==2,1,888)))))

# Labels mapping
label_map <- c(
  Non = 0,
  Oui = 1,
  `Ne sait pas` = 888
)

# Apply labels using mutate() and across()
WFP_Chad_MAD <- WFP_Chad_MAD %>%
  mutate(across(c(EverBreastF,PCIYCBreastF,PCMADStapCer,PCMADVegOrg,PCMADStapRoo,PCMADVegGre,PCMADFruitOrg,PCMADVegFruitOth,PCMADPrMeatO,PCMADPrMeatF,PCMADPrEgg,PCMADPrFish,PCMADPulse,PCMADDairy,PCMADFatRpalm,PCMADSnfChild,PCMADSnfPowd,PCMADSnfLns),
                ~ labelled::labelled(., label_map)))

```




```{r}


write_sav(WFP_Chad_MAD, paste0(dir_output_data,"/WFP_Chad_MAD.sav"))
write_dta(WFP_Chad_MAD, paste0(dir_output_data,"/WFP_Chad_MAD.dta"))
```



```{r eval=FALSE, include=FALSE}




#Creat Minimum Dietary Diversity 6-23 months (MDD)
# for population assesments - SNF is counted in cereals group (MDD)
# for WFP programme monitoring - SNF is counted in meats group (MDD_wfp)

#this version is for population assessments - SNF is counted in cereals group
data <- data %>% mutate(
                        MAD_BreastMilk = case_when( 
                          #PCMADBreastfeed == 1 ~ 1, TRUE ~ 0),
                        MAD_PWMDDWStapCer = case_when( 
                          PCMADStapCer == 1 | PCMADStapRoo == 1  | PCMADSnf == 1 ~ 1, TRUE ~ 0),
                        MAD_PulsesNutsSeeds = case_when( 
                          PCMADPulse == 1 ~ 1, TRUE ~ 0),
                        MAD_Dairy = case_when( 
                          # PCMADInfFormula == 1 | PCMADMilk == 1 | PCMADYogurtDrink == 1 | PCMADYogurt == 1 | PCMADCheese == 1 ~ 1,  TRUE ~ 0),
                        MAD_MeatFish = case_when( 
                          PCMADPrMeatO == 1 |  PCMADPrMeatPro == 1 | PCMADPrMeatF == 1 | PCMADPrFish == 1 ~ 1, TRUE ~ 0),  
                        MAD_Eggs = case_when( 
                          PCMADPrEgg == 1 ~ 1,  TRUE ~ 0),
                        MAD_VitA = case_when( 
                          PCMADVegOrg == 1 | PCMADVegGre == 1 | PCMADFruitOrg == 1 ~ 1, TRUE ~ 0),
                        MAD_OtherVegFruits = case_when( 
                          PCMADFruitOth == 1 | PCMADVegOth == 1 ~ 1, TRUE ~ 0)
                        )
#add together food groups to see how many food groups consumed
data <- data %>% mutate(MDD_score = MAD_BreastMilk +MAD_PWMDDWStapCer +MAD_PulsesNutsSeeds +MAD_Dairy +MAD_MeatFish +MAD_Eggs +MAD_VitA +MAD_OtherVegFruits)
#create MDD variable which records whether child consumed five or more food groups
data <- data %>% mutate(MDD = case_when(
                              MDD_score >= 5 ~ 1, TRUE ~ 0))

var_label(data$MDD) <- 'Minimum Dietary Diversity (MDD)'
val_lab(data$MDD) = num_lab("
             0 Does not meet MDD
             1 Meets MDD
")

##this version for WFP programme monitoring - SNF is counted in meats group
data <- data %>% mutate(
  MAD_BreastMilk_wfp = case_when( 
    PCMADBreastfeed == 1 ~ 1, TRUE ~ 0),
  MAD_PWMDDWStapCer_wfp = case_when( 
    PCMADStapCer == 1 | PCMADStapRoo == 1 ~ 1, TRUE ~ 0),
  MAD_PulsesNutsSeeds_wfp = case_when( 
    PCMADPulse == 1 ~ 1, TRUE ~ 0),
  MAD_Dairy_wfp = case_when( 
    PCMADInfFormula == 1 | PCMADMilk == 1 | PCMADYogurtDrink == 1 | PCMADYogurt == 1 | PCMADCheese == 1 ~ 1,  TRUE ~ 0),
  MAD_MeatFish_wfp = case_when( 
    PCMADPrMeatO == 1 |  PCMADPrMeatPro == 1 | PCMADPrMeatF == 1 | PCMADPrFish == 1 | PCMADSnf == 1 ~ 1, TRUE ~ 0),  
  MAD_Eggs_wfp = case_when( 
    PCMADPrEgg == 1 ~ 1,  TRUE ~ 0),
  MAD_VitA_wfp = case_when( 
    PCMADVegOrg == 1 | PCMADVegGre == 1 | PCMADFruitOrg == 1 ~ 1, TRUE ~ 0),
  MAD_OtherVegFruits_wfp = case_when( 
    PCMADFruitOth == 1 | PCMADVegOth == 1 ~ 1, TRUE ~ 0)
)
#add together food groups to see how many food groups consumed
data <- data %>% mutate(MDD_score_wfp = MAD_BreastMilk_wfp +MAD_PWMDDWStapCer_wfp +MAD_PulsesNutsSeeds_wfp +MAD_Dairy_wfp +MAD_MeatFish_wfp +MAD_Eggs_wfp +MAD_VitA_wfp +MAD_OtherVegFruits_wfp)
#create MDD variable which records whether child consumed five or more food groups
data <- data %>% mutate(MDD_wfp = case_when(
  MDD_score_wfp >= 5 ~ 1, TRUE ~ 0))

var_label(data$MDD_wfp) <- 'Minimum Dietary Diversity for WFP program monitoring (MDD)'
val_lab(data$MDD_wfp) = num_lab("
             0 Does not meet MDD
             1 Meets MDD 
")


table(data$MDD)
table(data$MDD_wfp)

#Calculate Minimum Meal Frequency 6-23 months (MMF) 

#Recode into new variables turning dont know and missing valules into 0  - this makes syntax for calculation simpler

data <- data %>% mutate(PCMADBreastfeed_yn = case_when(
                                            PCMADBreastfeed == 1 ~ 1, TRUE ~ 0),
                        PCMADMeals_r = case_when(
                                            between(PCMADMeals,1,7) == 1 ~ PCMADMeals, TRUE ~ 0),
                        PCMADInfFormulaNum_r = case_when(
                                            between(PCMADInfFormulaNum,1,7) == 1 ~ PCMADInfFormulaNum, TRUE ~ 0),
                        PCMADMilkNum_r = case_when(
                                            between(PCMADMilkNum,1,7) == 1 ~ PCMADMilkNum, TRUE ~ 0),
                        PCMADYogurtDrinkNum_r = case_when(
                                            between(PCMADYogurtDrinkNum,1,7) == 1 ~ PCMADYogurtDrinkNum, TRUE ~ 0)
                        )

data <- data %>% mutate(MMF = case_when(
   PCMADBreastfeed_yn == 1 & between(PCMADChildAge_months,6,8) & (PCMADMeals_r >= 2)  ~ 1, #for child breast fed yesterday - between 6 - 8 months minimum number of meals is 2
   PCMADBreastfeed_yn == 1 & between(PCMADChildAge_months,9,23) & (PCMADMeals_r >= 3) ~ 1, #for child breast fed yesterday - for children above 8 months minimum number of meals is 3
   #for child non breast fed regardless of age , child must have had 1 or more meal of solid/semisolid/soft food and child had 4 or more liquid + solid/semisolid/soft meals 
   PCMADBreastfeed_yn == 0 & (PCMADMeals_r >= 1) & (PCMADMeals_r + PCMADInfFormulaNum_r + PCMADMilkNum_r + PCMADYogurtDrinkNum_r >= 4) ~ 1,
   TRUE ~ 0)
)

var_label(data$MMF) <- 'Minimum Meal Frequency (MMF)'
val_lab(data$MMF) = num_lab("
             0 Does not meet MMF
             1 Meets MMF
")

#Calculate Minimum Milk Feeding Frequency for non-breastfed children 6-23 months (MMFF) 

data <- data %>% mutate(MMFF = case_when(
                              PCMADBreastfeed_yn == 0 & (PCMADInfFormulaNum_r + PCMADMilkNum_r + PCMADYogurtDrinkNum_r >= 2) ~ 1,
                              TRUE ~ 0)
)
                    
var_label(data$MMFF) <- 'Minimum Milk Feeding Frequency for non-breastfed children (MMFF)'
val_lab(data$MMFF) = num_lab("
             0 Does not meet MMFF
             1 Meets MMFF
")


# Minimum Acceptable Diet (MAD)

#For breastfed infants: if MDD and MMF are both achieved, then MAD is achieved
#For nonbreastfed infants: if MDD, MMF and MMFF are all achieved, then MAD is achieved


#using MDD for population assesments
data <- data %>% mutate(MAD = case_when(
                        PCMADBreastfeed_yn == 1 & MDD == 1 & MMF == 1 ~ 1,
                        PCMADBreastfeed_yn == 0 & MDD == 1 & MMF == 1 & MMFF == 1 ~ 1,
                        TRUE ~ 0))

#using MDD for WFP program monitoring
data <- data %>% mutate(MAD = case_when(
  PCMADBreastfeed_yn == 1 & MDD_wfp == 1 & MMF == 1 ~ 1,
  PCMADBreastfeed_yn == 0 & MDD_wfp == 1 & MMF == 1 & MMFF == 1 ~ 1,
  TRUE ~ 0))

var_label(data$MAD) <- 'Minimum Acceptable Diet (MAD)'
val_lab(data$MAD) = num_lab("
             0 Does not meet MAD
             1 Meets MAD
")


#creates a table of the weighted percentage of MAD
#and providing the option to use weights if needed

MAD_table_wide <- data %>% 
  drop_na(MAD) %>%
  count(MAD) %>% # if weights are needed use instead the row below 
  #count(MAD, wt = nameofweightvariable)
  mutate(Percentage = 100 * n / sum(n)) %>%
  ungroup() %>% select(-n) %>%
  pivot_wider(names_from = MAD,
              values_from = Percentage,
              values_fill =  0) 

```
